Splitting Space and IdGen; resolve is now an abstract method of Space
