+++ src/omega/space.jl
@@ -1,33 +1,13 @@
 module Space
 
-using ..Misc
-
-export Ω, ΩWOW, uid, @id,
-       Id, Ints
+export Ω, ΩWOW, resolve
 
 "Probability Space indexed with values of type I"
 abstract type Ω{I} <: AbstractRNG end
 
+function resolve end
+
 abstract type ΩWOW{I} <: Ω{I} end
-
-const uidcounter = Counter(0)
-
-"Unique id"
-uid() = (global uidcounter; increment(uidcounter))
-
-"Construct globally unique id for indices for ω"
-macro id()
-  uid()
-end
-
-"Index of Probability Space"
-const Id = Int
-
-"Tuple of Ints"
-const Ints = NTuple{N, Int} where N
-
-"Id of a random variable"
-const RandVarId = Int
 
 ## Rand
 ## ====
